92% faster:

Before:
12:45.96s real, 720.74s user, 27.77s sys, 3180288Kb RSS

After
6:38.35s real, 375.40s user, 15.98s sys, 3423064Kb RSS