@boxdot 2018-11-28
Interesting. I guess this is because of contiguous access pattern of the strings.