@VeaaC 2018-11-28
No, this is because the string lookup in PBF is deduplicated. Previously we tried to insert the same string multiple times (once for each usage).