@boxdot 2018-11-28
Got it. It is dedup per block. My implementation did not rely on this assumption. Probably, we should state the assumption somewhere in comments.